fix(typdoc): add typedoc.jsosn and use library mode
