+++ package.json
@@ -4,7 +4,7 @@
   "description": " A browser-compatible JSON-RPC client with multiple transports.",
   "main": "build/index.js",
   "scripts": {
-    "build": "tsc && typedoc --out docs --includeDeclarations --excludeExternals  --excludeNotExported --excludePrivate --hideGenerator --mode file --readme none",
+    "build": "tsc && typedoc src/index.ts",
     "lint": "tslint --fix -p .",
     "test": "npm run lint && jest --coverage"
   },
@@ -29,7 +29,7 @@
     "serve-handler": "^6.1.3",
     "ts-jest": "^25.0.0",
     "tslint": "^6.0.0",
-    "typedoc": "^0.17.0",
+    "typedoc": "^0.17.0-3",
     "typescript": "^3.4.5"
   },
   "dependencies": {
