+++ typedoc.json
@@ -1,0 +1,5 @@
+{
+  "name": "@open-rpc/client-js",
+  "out": "docs",
+  "mode": "library"
+}
